+++ hub_module/modules/video/classification/VideoTag_TSN_AttentionLSTM/README.md
@@ -1,10 +1,10 @@
 ```shell
-$ hub install videotag_tsn_lstm==1.0.0
+$ hub install VideoTag_TSN_AttentionLSTM==1.0.0
 ```
 
 ## 命令行预测示例
 ```shell
-hub run videotag_tsn_lstm --input_path 1.mp4 --use_gpu False
+hub run VideoTag_TSN_AttentionLSTM --input_path 1.mp4 --use_gpu False
 ```
 **参数**
 示例文件下载：
@@ -40,7 +40,7 @@ def classification(paths,
 ```python
 import paddlehub as hub
 
-videotag = hub.Module(name="videotag_tsn_lstm")
+videotag = hub.Module(name="VideoTag_TSN_AttentionLSTM")
 
 # execute predict and print the result
 results = videotag.classification(paths=["1.mp4","2.mp4"], use_gpu=True, save_dir="predict_dir")
