+++ hub_module/modules/video/classification/VideoTag_TSN_AttentionLSTM/resource/configs/attention_lstm.yaml
