+++ hub_module/modules/video/classification/VideoTag_TSN_AttentionLSTM/resource/models/attention_lstm/README.md
