+++ hub_module/modules/video/classification/VideoTag_TSN_AttentionLSTM/resource/models/tsn/tsn.py
